Changed Liquid to store as a Type rather than Int
This is a major change in how everything is stored and is only first
step. This change water being ID I to Liquid.WATER. Makes it so much
easier to understand what a liquid is and carries extra info like the
block that is linked to the liquid. The Name of the liquid.

I also started on a way to stored beams by create a class like Clac's
Vectors.  It stores a beam as a new beam with intensity, canProduce
light,and direction moving.