worked on mechanical system and updated for ue8.7
Simi unstable due to how i've been messing around with mechanical
systems, but it looks good enough. It also seems to work fine on UE8.7,
and i added dependence checks to remove some error on load. On top of
that i update version number for both mods to be the same. Will help
track versions better