bug fixes and packet reductions
fixed another bug with the pipe, pipe would not lose pressure after
being disconncted from source. Also cleaned up a bit on some of the code
and added code to only send packets if the machine updated. This will
reduce a lot of paceket load being sent to the client.