Correct Slanted Belt Render Issue
Some one on the forum noticed this issue with Slanted Belts not
rendering connected. I didn't know this was a feature that was usable or
i would have caught it sooner. Anyways its mostly fixed but needs a
texture which i'll try to create.

Also who every made the conveyor belt textures animated like that nice
job.