Fixed a SMP crash issue with sound for belts
Not sure if you guys are even bug testing on SMP but you need too every
so often. SMP lan, and SMP Dedicated are two diffrent server types. The
SMP lan can get away with something like playing sound. However, doing
anything client related on SMP Dedicated will cause a crash. Make sure
in the code anything client releated is only done on the client reguard
less of how small.

Also who every changed the rejector to use the paper item and display
the item name above it good job. Its a nice feature and better than the
GUI I  had, however if there isn't a way alread we need to work in a way
to have several item on a rejector. Which we can either do with the
detector, another block, or the rejector itself.

On another note i turned down the AMP pull a little to make belt lines
easier to feed.