sync check
I think either my forge or mcp is not the same version as you guys since
right off the back several of the var were diffrent. Might want to
update your forge and check.

Another note i'm starting to work on our crane design as a way to make
the quarry as well. I'll post my though on it in the issue section for
us to talk about.