Item Crate can add items to Block Crate
If you have a crate, with items in it, clicking on a create that is
empty or contains the same item will result in the held crate giving
some of its items to the block crate. So far now bugs but futhur testing
will be need to make sure that the max limit is held as well as the item
crate can still be used.

Also not sure why the two computer craft api files think there new.