Done with crane model
Ok i'm done but non of the textures sheets are lined up. I'll do this
later or a texture artist is welcome to line the sheet up.

CraneArm is the arm this will be rendered where every the target is at.
There is a block in it that needs to be removed from the code when
added. IT is there just to help me model it. If anyone knows how it can
also be used to render the block texture on when the arm picks up a
block

craneArmHoist is were the arm connects to the craneMount point. Metal
cables need to be rendered from it to the craneArm. Look at Beacon code
for help or use Calclavia's Laser Fx with some tweaks.