Fixed a pipe error plus changed some functions
This is were bug testers are so nice. Turns out that machine from other
mods could directly put the wrong liquid into the pipe the way i had it
coded.

Also i'm going to work threw my api and clean it up the best i can. A
few of the functions i have are unneed as well as not very well made.