merged IPsiMachine with IFluidPipe
They are now known as ILiquidNetworkPart since any tileEntity can act as
a conveyor of liquids in the network as long as they use that interface
and its methods right. This allows for pipes, as well as guages, fluid
motors, and other things to exist in the network without its spliting or
changing the over all network.