reworked the network code
added comments and change comments in several places
added can connect method to psi receiver and change the receiver method
to get a liquid stack so it can pass that stack to the next network.
changed onOverCharge to onPressureChange that will check each time the
pressure network increases if one of the parts will bust.