Created a new networkPipe base on UE's Conductor
My network code is about the same as UE's short of the fact i have to
still ably to net.minecraftForge.Liquids interfaces in order to keep my
mod compatibly with other mods. So its only logical that my network pipe
updates, reacts, and works the same as the UE conductor.  Though i will
be making change to mine in order to have it function with liquids more
and more.