setting up tank for IConnectionProvider
for now I will just use it for connection rules for rendering, and trade
processing. However, in the future i might use this with the path finder
to make a more uniform movement of liquids in the connected tanks. Might
even have them update all as one entity instead of each one at a time.
Though that will take some time if i set it up just like the UE wire's
network. As i will have to make sure that the liquid always travels to
the lowest tank first, and saves to the lower tanks. Which means that
i'll end up storing the tank's Y pos or getting it during
saving/loading/processing.