Merged pathfinder into one class
This path finder no longer depends on Calc's PathFinder to function.
however it is derived from his work so he still gets credit.

The new pathfinder works about the same as calc's except instead of
trying to find a path its trying to find all resources along the path. I
still need to improve this to have a dumped down version of A* so that
i'm filling the closest blocks first and draining the furthest block
first.