Added more Mechanical api stuff
Not sure if i want to work on this right now as its just a bonus
addition to FluidMech a this point. Though it badly needs overhauled to
include its own pathfinder. The rods need to work as one rod using the
same pos, spin speed, Force, etc. As well math needs to be added to calc
force correctly