Fixed generic pipe connections
Was doing the connection checks wrong. Now its check for meta data if
the pipe uses the same class file for connections. Its should connection
to any generic pipe of its color and iron pipes. This should allow for
unique pipe networks for generic liquid types. This however will not
allow for a set liquid type filtering or transport.

Also in an earlier push i changed generic pipes to only move 1 bucket a
request but 3 request for non-generic pipes a request.