Disabled Hopper interaction due to bugs
I can't explain what is going on between the crate and hopper. As far as
i can see the crate works fine with pipes, and other things. However the
hopper creates new bugs that i can't solve at the moment. It will keep
trying to input items into the crate after the crate is full. It will
then after world load trash a crates inv that its pulling from. Since i
believe that the hopper is the root cause i'm going to just disable it
use with crates.