Created a crafting manager to reduce code per class
The imprinter class was starting to get messy with all of its other
non-crafting code. To help make things cleaner and to create a reusable
class i have moved the crafting code to its own class. To use the class
you simple need to have the TileEntity implement IInventory(or have an
internal Inv refrence) and implement IAutoCrafting. Also i'll be moving
these classes to Dark-Lib when done.