Massive Fluid Network changes
Got started on this and didn't stop to better document the changes.
First the man changes are to div out the network code to be more usable
in other way than fluids. As well to make a new network class for the
tanks which are now becoming a more like the pipes. I will later as well
make a mechanical network and a data network for later use.

On top of that there are some helper class that i didn't upload before i
started work. There is nothing too them as they were in the start of
development.