worked on packet handling and power
changed the packet handler to use a method allowing for super classes to
work with the packet. This is without losing the packet ID or needing to
reread the data from the packet again.

As well more improvement to how the tile understands how its powered
from the network. Added ways to detect if the tile was remove, no longer
connected, and a delay in case the power source missed a power cycle.
More improvement are to come especial with how much power is used by the
entire network.