more work on assembly network
still buggy as in not working past powering assembly tiles. As well it
seems to be getting a 100W packet of power every 2 seconds from both
mekanism and buildcraft power pipes. It should not be storing anymore
than the internal battery size but its seem to be anyways. As well max
packet size should only be equal to the request which as well is not the
case.