Merged TheDarkMachine into core
At first i was going to keep the core as a simple loader but seeing that
need for content out of my DarkMachine mod i merged the two. This will
cause a bit of extra stuff for the player to deal with. However, i did
add configs to disable the ore and ore Items from the mod.