Changed start pump to drain multi sources
I've been avoiding this because i wanted to remove the starter pump.
However, seeing as early game it is needed i wanted to rework it. As
well since finite fluids are become more and more common i changed the
code. It will now drain sources within 20 blocks of itself. I'll still
need to redo the path finder, render, and several other parts to match
the change i want. I'm thinking of doing a buildcraft style pipe
extension from the base. However, with a full render and animation to go
with it. The other option that came to mind was a hose attachment
allowing it to be a basic version of the construction pump without the
drain.