swapped Balance Tank method for read/write tiles method
Honestly shouldn't act to different from how its was seeing as nothing
really changed. Only thing that happened was that the load and write
part of BalanceCollectiveTank was split into two different methods. Will
need some testing though to check if everything works.

Also for NetworkFluidContainers its balance method was changed a bit to
maybe fix errors with fluid remaining in the tank that shouldn't have
fluid in them. Render will still need to be changed to fix this a little
better.