Toy with battery box model design
Not really liking this version and may have to work on it some more. The
idea is to create a way to insert 4 batteries into the machine. Then for
these batteries to visually show on the out side of the machine with
there charging info. As well placing two machine together will need to
effect the render.