Fixed create loading items that don't exist
Since the create saves item stacks different than MC it would load them
even if the item they went too no longer existed. I'll work on a more
improved fix for this that will save the unlocalized name so in case the
IDs change the create remembers what was in it.