Reverted new network changes
Something wasn't going write and right now is not the time to create new
bugs. So for now i'll work with the old network and improve it first.
Then i'll create a new network separated from the old one for multi-tank
setup. When it get to the point that can do the job of the old one and
the new one i'll remove the old network.