Started working on a complex Fluid & Object mixing system
Too start with i original was avoiding making such a system since it
boils down to an object crafting system. Though with my future and
current plans i decided to start working on it.

The system, when finish, will be able to handle mixing of fluids, Items,
Blocks, and anything to form a result of some kind. The idea is to used
it in the pipes, tanks, brewing stand, boilers, and other machines to
automated the process of mixing fluids. For pipes it will either create
a new fluid, or prevent the pipe from working. Tanks will do the same
but can become filled with blocks or items. Brewing stand will use the
system to create new fluids, gases, potions, foods, drinks, items, and
so on. Boiler will use the simple version of it to convert a fluid to
another fluid without any other processing.