Created a crusher machine
First machine in a series of machines. This is designed to be a low
capacity machine with a single stack input. Larger version will be added
later that can process items as fast as they are inputted. As well it
will be belt based, and will not actually store the item.  Instead it
will process the item on the belt and does have a chance to miss the
item if the belt goes too fast.