reworked how ores are setup
Removed the dependency on using the material enum as it was adding ores
that should not generate which was causing a lot of issues. So to solve
this the block has its own enum set of ores it will work with. However,
i still need to reset the ore generator