Toyed with a new way to register TileEntitySpecialRenderers
The idea is to let the object being registered control all aspects of
how its registered. This started with the blocks adding config files,
disabling themselves, and loading their own tile entities. Now i'm
working on having the tile entities register their own renderers, and
have configs as well. Give that the registration is actually handled by
the ModObjectRegistry and the object just provided the data instead of
the mod class.