Started work on a enum to contain element data
Plan is to include all elements from the periodic table with as much
information as possible per elements. This might get long so i might
want to consider constructing a data base that will be parsed per use.
That way only a few elements are loaded into memory instead of over 255.