Filled out basic info for each element
Ctrl+v for the win. I honestly copied the data out of a csv file and
manually added the ( " , to each line. Honestly this was a stupid move
as i could code something to read and import that csv file. Though i'm
looking to create a code based version. I still need to at least add
density per element and then heating values for the ones i plan to use.