Worked a bit on a new terminal system
This is more of me starting back up on global access lists for sentry
guns and chests. The idea right now is to create a system like bukkit
plugin permissions has. In which a player can create a group, assign
command node, and then assign players. As well can assign direct command
nodes to the player user instance. These groups will be managed from a
GUI that can be access anytime. The data will be saved to a NBT file in
the map, and can be access by any object. Objects will not save a copy
of the list but will request it to be loaded.