Lot of work on armbot task system
Started designing the task system to work with other non-armbot devices.
As well fixed up several command, and designed a way for the command to
say something is wrong. Will continue to improve this before i actually
start on the encoder and armbot. That way there is less changes after
recoding both devices.