More work and some package changes
Changed back in api from armbot to coding since it no longer just
applies to the armbot. There are plans to move this to the core and
start adding coding to all parts of the mod.

Added a way for the encoder to understand limits on arguments. This way
it can clearly show the limits to the user.