Toyed with the idea of XML help pages
Its a very good idea but a bit harder than i though. I'm going to have
to get very creative with what i can pull from the game. Including how i
want to handle textures as they need to be loaded in a way both the mod
can use and later a website design. Though i want to avoid having the
xml file be packaged with textures inorder to work.