Worked on encoder
We have something though its not much. The encoder now has an inventory,
code, and help gui. The inventory gui needs a slot for the disk and
possible more slots for other disk. Possible should add a way to clear
the disk and clone to another disk. The code gui is not close to finish
though it now does render parts of the task. Still working on lining
things up and thinking about expanding the size. The help gui has
nothing.