Started up on redoing arch's gas system
Not that it was bad archadia but we have the fluid system which i'm
using. Though to help out a bit i took your gas and made it extend the
fluid class. This way you can create gas fluid faster. As well i started
to create an enum to store all the gases with extended element data.
Later on this will allow use to do advanced features like gas increasing
pressure when heated.