Changed the key event to stop when key is used
Forgot to add this when i created the handler. This will cut down a bit
of CPU time when keys are received though i might be able to cut this
down more by sending this with the entity id. That way it goes directly
to the entity that is going to use it.