Fixed most of the rendering issues
Turns out i was in the middle of converting AL to the registry system
and forgot to finish. So a few builds were pushed were no render code
was being register. This was due to how the registry system depends on
the block being registered to it to activate the other registry calls
for the block.