Removed test program from gui
well removed it from the gui but added it to the tile in the same exact
format. Mainly just to test that the gui can pull the program from the
tile. Once this test is done though the program will need to be moved to
server side. That way packet testing can begin. After that its on to
actual editing.