Starting to separate core from content
At some point in the near future the core will need to act as a library
more than a mod. Though it will still need to have its common items, and
blocks. Such as parts, and ores. Machines like the battery box, wire,
and generator will be moved to another mod.