Fixed inserter rotation being off by 3 degrees
Logic bug and not a rendering issue. The code was setup to only rotate until we hit an error bar of facing the direction. Issue was nothing forced the rotation to snap to facing. So it would be off by the error bar amount which was 3 degrees. 

This is sorta a temp fix so a better system would be needed. Something that snaps the rotation to the desired angle to prevent issues later on.